% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.eraser.R
\name{database.eraser}
\alias{database.eraser}
\title{Erases the data of all of selected oligonucleotides from a g4dbr database while maintaining its structure for reuse}
\usage{
database.eraser(
  db.to.erase = NULL,
  remove.oligos = NULL,
  erase.CD,
  erase.NMR,
  erase.MS,
  erase.UV
)
}
\arguments{
\item{db.to.erase}{An .Rda file path to erase}

\item{remove.oligos}{A vector containg strings of the \code{oligo} names}
}
\value{
Return a list of partially of fully emptied dataframes with the correct structure for g4db use
}
\description{
Erases the data of all of selected oligonucleotides from a g4dbr database while maintaining its structure for reuse
}
\examples{
database.eraser("this is a filepath")

}
