% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpsilonCalc.R
\name{epsilon.calculator}
\alias{epsilon.calculator}
\title{Calculates extinction coefficient from oligonucleotide sequence}
\usage{
epsilon.calculator(sequence = NULL)
}
\arguments{
\item{sequence}{A string.}
}
\value{
The extinction coefficient of \code{sequence}.
}
\description{
Calculates extinction coefficient from oligonucleotide sequence
}
\examples{
epsilon.calculator("GGGTTAGGGTTAGGGTTAGGG")
}
