% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{g4db}
\alias{g4db}
\title{G4 biophysics database management and visualization}
\usage{
g4db()
}
\value{
Launches the shiny app.
}
\description{
G4 biophysics database management and visualization
}
\examples{
g4db()
}
