% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massdiet.R
\name{mass.diet}
\alias{mass.diet}
\title{Mass spectrometry data size reduction}
\usage{
mass.diet(fat.mass, base.start, base.end, range.start, range.end, baseline.int)
}
\arguments{
\item{fat.mass}{A dataframe with mz, and int columns.}

\item{base.start}{A numeric value.}

\item{base.end}{A numeric value, larger than \code{base.start}.}

\item{range.start}{A numeric value.}

\item{range.end}{A numeric value, larger than \code{range.start}.}

\item{baseline.int}{A numeric value.}
}
\value{
Size-reduced MS data.
}
\description{
Mass spectrometry data size reduction
}
\examples{
mass.diet(fat.mass = my_MS_data, base.start = 1750, base.end = 2000,
range.start = 1000, range.end = 2500, baseline.int = 2)
}
